Warbler Setup Part 1: Fix current features Step 1: Understand model Step 2: Fix logout & other security problems Step 3: Fix user profile Step 4: Fix user cards Step 5: Profile edit Step 6: Fix homepage Step 7: Research & understand login strategy Part 2: Add likes Part 3: Add tests Solution Further study Custom 404 page Add AJAX DRY up the templates DRY up the authorization DRY up the URLs Optimize queries Make a change password form Allow “private” accounts Add admin users User blocking Direct messages Warbler» Download
A Twitter clone with a Rithm spin!

_images/warbler.png
This is a multi-day exercise to extend a somewhat-functioning Twitter clone. It is intended to give you practice reading and understanding an existing application, as well as fixing bugs in it, writing tests for it, and extending it with new features.

IMPORTANT Clean & Commit As You Go
Treat this exercise like a real code base:

tidy and document as you go
check into Git often, as you add new features
Practicing this process will help you as you work on larger codebases.

Also, you may find bugs not mentioned in the app as you explore it. Fix these, and keep a log of what you fixed.

Setup»
Create the Python virtual environment:

$ python3 -m venv venv
$ source venv/bin/activate
(venv) $ pip install -r requirements.txt
Set up the database:

(venv) $ psql
=# CREATE DATABASE warbler;
=# (control-d)
(venv) $ python seed.py
Create an .env file to hold configuration:

.env»
SECRET_KEY=abc123
DATABASE_URL=postgresql:///warbler
Start the server:

(venv) $ flask run
WARNING Running Flask on some Macs
If you get an error that your address is already in use after starting up flask with flask run, you may need to run on port 5001 instead. Some Macs running newer software already have a process running on port 5000 by default, so the port is already taken.

To run flask on port 5001, run flask run -p 5001 from the command line.

If you’re curious and would like to see what processes are listening on port 5000 at any given time, the command lsof -i :5000 will give you that info!

Part 1: Fix current features»
Step 1: Understand model»
Read models.py. Make a diagram of the three tables.

Note that the follows table has an unusual arrangement: it has two foreign keys to the same table. Why?

Using pdb, set a debugger in one of your routes, and hit it to pause code execution. From here, try out the various relationships set on the model classes. Get a feel for how they work.

Step 2: Fix logout & other security problems»
Right now, there is a link to logout in the base template, but we have a bug here and the logout route logic is not implemented.

Find and fix the bug, then implement the view function logic (make sure your implementation is secure). On logout, it should flash a success message and redirect to the login page.

There are other security problems in the starter code. Read through our starter code in app.py and look for problems like not authenticating properly or not handling CSRF attacks. Fix these.

Step 3: Fix user profile»
The profile page for users works, but is missing a few things:

the location
the bio
the header image (which should be a background at the top)
Add these.

Step 4: Fix user cards»
On the followers, following, and list-users pages, the user cards need to show the bio for the users. Add this.

Step 5: Profile edit»
There are buttons throughout the site for editing your profile, but this is unimplemented.

It should ensure a user is logged on (you can see how this is done in other routes)
It should show a form with the following:
username
email
image_url
header_image_url
bio
password [see below]
It should check that that password is the valid password for the user—if not, it should flash an error and re-present the form with the data entered so far reappearing.
It should edit the user for all of these fields except password (ie, this is not an area where users can change their passwords–the password is only for checking if it is the current correct password.
On success, it should redirect to the user detail page.
Step 6: Fix homepage»
The homepage for logged-in-users should show the last 100 warbles only from the users that the logged-in user is following, and that user, rather than warbles from all users.

Step 7: Research & understand login strategy»
Look over the code in app.py related to authentication.

How is the logged in user being kept track of?
What is Flask’s g object?
What is the purpose of add_user_to_g?clear

+++ app.py
@@ -36,6 +36,8 @@ def add_user_to_g():
     if CURR_USER_KEY in session:
         g.user = User.query.get(session[CURR_USER_KEY])
 
+        g.form = CsrfForm()
+
     else:
         g.user = None
 
@@ -125,11 +127,11 @@ def login():
 def logout():
     """Handle logout of user and redirect to homepage."""
 
-    form = CsrfForm()
+    if not g.user:
+        flash("Access unauthorized", "danger")
+        return redirect("/")
 
-    # TODO: add user check, refactor security check
-
-    if form.validate_on_submit():
+    if g.form.validate_on_submit():
         session.pop(CURR_USER_KEY)
         flash("You've been successfully logged out", "info")
         return redirect('/')
@@ -148,7 +150,7 @@ def list_users():
     Can take a 'q' param in querystring to search by that username.
     """
 
-    form = CsrfForm()
+    # form = CsrfForm()
 
     if not g.user:
         flash("Access unauthorized", "danger")
@@ -161,14 +163,14 @@ def list_users():
     else:
         users = User.query.filter(User.username.like(f"%{search}%")).all()
 
-    return render_template('users/index.html', users=users, form=form)
+    return render_template('users/index.html', users=users, form=g.form)
 
 
 @app.get('/users/<int:user_id>')
 def show_user(user_id):
     """Show user profile."""
 
-    form = CsrfForm()
+    # form = CsrfForm()
 
     if not g.user:
         flash("Access unauthorized", "danger")
@@ -176,35 +178,35 @@ def show_user(user_id):
 
     user = User.query.get_or_404(user_id)
 
-    return render_template('users/show.html', user=user, form=form)
+    return render_template('users/show.html', user=user, form=g.form)
 
 
 @app.get('/users/<int:user_id>/following')
 def show_following(user_id):
     """Show list of people this user is following."""
 
-    form = CsrfForm()
+    # form = CsrfForm()
 
     if not g.user:
         flash("Access unauthorized", "danger")
         return redirect("/")
 
     user = User.query.get_or_404(user_id)
-    return render_template('users/following.html', user=user, form=form)
+    return render_template('users/following.html', user=user, form=g.form)
 
 
 @app.get('/users/<int:user_id>/followers')
 def show_followers(user_id):
     """Show list of followers of this user."""
 
-    form = CsrfForm()
+    # form = CsrfForm()
 
     if not g.user:
         flash("Access unauthorized", "danger")
         return redirect("/")
 
     user = User.query.get_or_404(user_id)
-    return render_template('users/followers.html', user=user, form=form)
+    return render_template('users/followers.html', user=user, form=g.form)
 
 
 @app.post('/users/follow/<int:follow_id>')
@@ -253,35 +255,31 @@ def profile():
 
     user = User.query.get_or_404(g.user.id)
 
-    form = UserEditForm()
+    form = UserEditForm(obj=user)
 
     if form.validate_on_submit():
+
         user = User.authenticate(
             user.username,
             form.password.data,
         )
-        #TODO: prepopulating trims both routes \/
+
         if user:
-            user.username = form.username.data or user.username
-            user.email = form.email.data or user.email
-            user.image_url = form.image_url.data or user.image_url
-            user.header_image_url = form.header_image_url.data or user.header_image_url
-            user.bio = form.bio.data or user.bio
+            user.username = form.username.data
+            user.email = form.email.data
+            user.image_url = form.image_url.data
+            user.header_image_url = form.header_image_url.data
+            user.bio = form.bio.data
 
             db.session.commit()
             return redirect(f"/users/{user.id}")
 
         else:
-            form.username = form.username.data,
-            form.email = form.email.data,
-            form.image_url = form.image_url.data,
-            form.header_image_url = form.header_image_url.data,
-            form.bio = form.bio.data,
 
             flash("Incorrect password", "danger")
             return render_template('users/edit.html', form=form)
 
-    return render_template('/users/edit.html', form=form)
+    return render_template('/users/edit.html', form=form, user=user)
 
 
 @app.post('/users/delete')
@@ -290,18 +288,19 @@ def delete_user():
 
     Redirect to signup page.
     """
-    #TODO: security check
 
     if not g.user:
         flash("Access unauthorized.", "danger")
         return redirect("/")
 
-    do_logout()
+    if g.form.validate_on_submit():
 
-    db.session.delete(g.user)
-    db.session.commit()
+        do_logout()
 
-    return redirect("/signup")
+        db.session.delete(g.user)
+        db.session.commit()
+
+        return redirect("/signup")
 
 
 ##############################################################################
